requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.9
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
Pillow==8.4.0
promise==2.3
psycopg2==2.9.2
pydotplus==2.0.2
pygraphviz==1.7
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
